ifit/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
numpy
mediapipe
opencv-python